+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 - Relative file input paths are converted to absolute paths.
 - New history-filter option added to define paths to be excluded from the history log feature.
 - New command to move the Window to the screen edge (Alt+Arrow) or center (Alt+BS).
+- Smaller OSD media info font size and more duplicated and obvious info removed from titles.
 - libmpv shinchiro 2022-08-11
 
 input.conf changes: