+++ docs/Changelog.md
@@ -6,8 +6,10 @@
 - New history-filter option added to define paths to be excluded from the history log feature.
 - New command to move the Window to the screen edge (Alt+Arrow) or center (Alt+BS).
 - Smaller OSD media info font size and more duplicated and obvious info removed from titles.
-- Improved mouse cursor hide behavior.
-- libmpv shinchiro 2022-08-11
+- Improved mouse cursor auto hide behavior.
+- Fix severe bug causing termination before scripts having a chance reacting to shutdown event.
+- libmpv shinchiro 2022-10-02, fixes a severe bug causing hangs:
+  https://github.com/mpv-player/mpv/pull/10569
 
 input.conf changes: